[#124] SonarQube 적용 및 리팩토링

+++ src/main/java/io/driver/codrive/global/util/DateUtils.java
@@ -58,13 +58,13 @@ public class DateUtils {
 
 	public LocalDateTime getMondayDateTime(LocalDate pivotDate) {
 		int pivotDay = pivotDate.getDayOfWeek().getValue();
-		int monday = DayOfWeek.MONDAY.getValue();
+		long monday = DayOfWeek.MONDAY.getValue();
 		return pivotDate.minusDays(pivotDay - monday).atStartOfDay();
 	}
 
 	public LocalDateTime getSundayDateTime(LocalDate pivotDate) {
 		int pivotDay = pivotDate.getDayOfWeek().getValue();
-		int sunday = DayOfWeek.SUNDAY.getValue();
+		long sunday = DayOfWeek.SUNDAY.getValue();
 		return pivotDate.plusDays(sunday - pivotDay).atTime(23, 59, 59);
 	}
 }
